chore: move PrettyPrintRecordKey to sn_protocol
